make lockin a unittest, not integration test
